# server/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Messenger API
  description: API мессенджера для личных чатов, групп и каналов
  version: 1.0.0
servers:
  - url: http://localhost:8888/api
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      description: Создает нового пользователя в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации данных
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      description: Авторизует пользователя и возвращает JWT токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверные учетные данные
...